Add App component, fix components paths
